Fix the process being leaked for real this time
I thought I had fixed this in the previous commit, turns out hidden
windows apparently don't receive the WM_CLOSE message in general.
I was temporarily under the impression that I'd fixed the issue but
I think I just hadn't hid the window when testing the last commit and
that resulted in the process closing correctly.

The way I solved this now is by grabbing a handle to a Lunar Magic
window that's normally used for issuing requests to Lunar Magic and
monitoring it. If the window is closed we know that Lunar Magic just
got closed and thus close our process too.

This is a pretty scuffed solution, but I don't think there's much else
I can do and it seems to work fine so I'm ok with it for now.

Fixes #7