Make Logger keep up with ROM changes and config errors.
Logger used to log to standard log file in directory which was first
opened, no matter what the current ROM directory was, if there was no
config file. Now it will check whether there's a config file in the
current ROM diretory:

- If there is no config file nothing will be logged (log level is set
to LogLevel::Silent)
- If there is a config file it will try to parse it
- If the parsing succeeds and log variables were specified they will
be used by the Logger
- If the parsing fails or the log variables aren't specified, the
default values will be used for logging (log_path will be set to
log to a file with the name "lunar_monitor_log.txt" in the **current**
directory and log_level will be set to LogLevel::Log)

The log messages about switching from ROM to ROM now include the ROM's
name for extra clarity.

The Logger no longer outputs a "Start of log" message on startup, but
rather whenever we switch to a new ROM, since that's the "start" of
logging for that particular ROM's directory.