Make LM pop up error message when saving shared palettes to moved ROM
Error message when saving shared palettes to ROM that is no longer there used to not pop up because the call to LMSaveSharedPalettesFunction was overwriting eax, which for some reason contains the HWND for MessageBoxA in LM's function for saving shared palettes and needs to be preserved.

Atari says the way I'm doing this is undefined behavior so I guess no guarantees that this will work forever/anywhere else.