Add button for exporting everything to LM and deprecate mwl_file_format
The button will export map16, all (edited) levels, global data and shared palettes to the locations specified by the valid config file in the current ROM's directory when clicked. If there is no (valid) config file found, the button will be greyed out.

This update also deprecates the `mwl_file_format` config variable, partially because it interferes with LM's `-ExportMultLevels` command line option (can't specify the same format there as for single levels) and partially because it's not a very useful option to begin with. `level #.mwl` will now always automatically be used as the format. Using `mwl_file_format` in a config file will raise an error (for now).

Closes #13, closes #12