Update Lunar Helper build report when exporting resources
This should hopefully work fine, I had some crashes at some point because frankly directory monitoring in win32 seems hellish, still not sure if it's fine now, might be leaking some handles or something, but I haven't gotten any crashes yet.

Updating the build report helps Lunar Helper not reinsert things that were already "inserted" by being saved to the ROM. 

What complicates this is that Lunar Monitor will reload the ROM if it detects a write to the build report, which would make it reload the ROM when it itself writes to the report. For now I circumvented this by only reloading the ROM if the `rom_hash` value in the build report changes, indicating that it was edited by Lunar Helper rather than Lunar Monitor.