Use last build time as reload criterium instead of ROM hash
ROM hash was sometimes an unreliable criterium  when switching back to state when previous ROM was built, because the hash of the ROM might stay the same when building, but the ROM we currently have may have changes that would have actually lead to a different hash.

Instead I'm using build time now, which should be more reasonable, note that it literally just checks if the time string is different, not if the actual time contained within it is different, but that shouldn't really matter since the only time it's ever changed is during a build.