Mark ROM as volatile if any resource export fails
If an export fails, the affected resource was still inserted into the ROM, thus the contents of the ROM are now volatile and should not be overwritten. Thus Lunar Monitor will now mark the ROM as volatile via the comment field if a resource *save* succeeded but Lunar Monitor failed to export it.